predict params, return cached file on download

+++ htx/utils.py
@@ -40,6 +40,8 @@ def download(url, output_dir, filename=None):
     if filename is None:
         filename = hashlib.md5(url.encode()).hexdigest()
     filepath = os.path.join(output_dir, filename)
+    if os.path.exists(filepath):
+        return filepath
     r = requests.get(url)
     r.raise_for_status()
     with io.open(filepath, mode='wb') as fout:
